requirements.txt
fastapi
uvicorn
moviepy
librosa
openai-whisper
mediapipe
numpy
pydantic
python-multipart